Removed the console.log() and added stock data into constants folder
