+++ frontend/src/constants/Carousel_data.jsx
@@ -1,0 +1,47 @@
+ import news from '../../../frontend/assets/Images/news.png';
+ 
+ const slides = [
+    {
+      id: 1,
+      image: news, 
+      title: 'INDIANS GO TO COMIC CON ?',
+      subtitle: 'ANUJ GOES TO COMIC CON DRESS AS',
+      description: 'Anuj goes to comic con dressed as Pablo Picasso!!\nYou thought he was a great designer ? Well he is,\nand more! He showcased his unique blah blah!'
+    },
+    {
+      id: 2,
+      image: news, 
+      title: 'INDIANS GO TO COMIC CON ?',
+      subtitle: 'ANUJ GOES TO COMIC CON DRESS AS',
+      description: 'Anuj goes to comic con dressed as Pablo Picasso!!\nYou thought he was a great designer ? Well he is,\nand more! He showcased his unique blah blah!'
+      
+      
+    },
+    {
+      id: 3,
+      image: news,
+      title: 'INDIANS GO TO COMIC CON ?',
+      subtitle: 'ANUJ GOES TO COMIC CON DRESS AS',
+      description: 'Anuj goes to comic con dressed as Pablo Picasso!!\nYou thought he was a great designer ? Well he is,\nand more! He showcased his unique blah blah!'
+      
+      
+    },
+    {
+      id: 4,
+      image: news, 
+      title: 'INDIANS GO TO COMIC CON ?',
+      subtitle: 'ANUJ GOES TO COMIC CON DRESS AS',
+      description: 'Anuj goes to comic con dressed as Pablo Picasso!!\nYou thought he was a great designer ? Well he is,\nand more! He showcased his unique blah blah!'
+      
+    },
+    {
+      id: 5,
+      image: news, 
+      title: 'INDIANS GO TO COMIC CON ?',
+      subtitle: 'ANUJ GOES TO COMIC CON DRESS AS',
+      description: 'Anuj goes to comic con dressed as Pablo Picasso!!\nYou thought he was a great designer ? Well he is,\nand more! He showcased his unique blah blah!'
+      
+    }
+  ];
+
+export default slides;
